fix: User DTOs refactor
