+++ Moodiverse/Moodiverse.WebAPI/Models/DTO/UserDTO/DeleteUserDto.cs
@@ -1,0 +1,6 @@
+namespace Moodiverse.WebAPI.Models.DTO.UserDTO;
+
+public class DeleteUserDto
+{
+    public string Password { get; set; }
+}
