+++ Moodiverse/Moodiverse.WebAPI/Models/DTO/UserDTO/LoginUserDto.cs
@@ -1,0 +1,7 @@
+namespace Moodiverse.WebAPI.Models.DTO.UserDTO;
+
+public class LoginUserDto
+{
+    public string Username { get; set; }
+    public string Password { get; set; }
+}
