+++ Moodiverse/Moodiverse.WebAPI/Models/DTO/UserDTO/UpdateUserDto.cs
@@ -1,11 +1,9 @@
-namespace Moodiverse.WebAPI.Models.DTO;
+namespace Moodiverse.WebAPI.Models.DTO.UserDTO;
 
 public class UpdateUserDto
 {
     public string Username { get; set; }
-    public string Email { get; set; }
-    public string Password { get; set; }
     public string FirstName { get; set; }
     public string LastName { get; set; }
-    public int SexId { get; set; }
+    public string Sex { get; set; }
 }
