+++ Moodiverse/Moodiverse.WebAPI/Models/DTO/UserDTO/UpdateUserEmailDto.cs
@@ -1,0 +1,6 @@
+namespace Moodiverse.WebAPI.Models.DTO.UserDTO;
+
+public class UpdateUserEmailDto
+{
+    public string NewEmail { get; set; }
+}
