+++ Moodiverse/Moodiverse.WebAPI/Models/DTO/UserDTO/UpdateUserPassDto.cs
@@ -1,0 +1,7 @@
+namespace Moodiverse.WebAPI.Models.DTO.UserDTO;
+
+public class UpdateUserPassDto
+{
+    public string OldPassword { get; set; }
+    public string NewPassword { get; set; }
+}
